requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
langchain==0.0.350
langchain-community==0.0.6
langsmith==0.0.69
pymilvus==2.3.3
openai==1.6.1
duckduckgo-search==3.9.9
grpcio